fix: use z.number() instead of z.number().int() in QuizItem
Anthropic structured output rejects the JSON Schema 'integer' type that
z.number().int() generates. Plain z.number() produces 'number' type which
is fully supported. Verified via Playwright — lesson generation succeeds.

+++ lib/ai/schemas.ts
@@ -4,7 +4,7 @@ import type { Domain } from '@/lib/domains'
 const QuizItem = z.object({
   question:    z.string(),
   options:     z.array(z.string()),   // Anthropic structured output doesn't support minItems > 1
-  answer:      z.number().int(),      // prompt instructs 0-3; .min/.max also rejected
+  answer:      z.number(),             // Anthropic doesn't support integer type or min/max
   explanation: z.string(),
 })
